refactor: split indexing module
