+++ agentic_index_cli/rank.py
@@ -2,7 +2,7 @@
 
 from typing import Dict
 
-from .agentic_index import compute_score as _compute_score
+from .scoring import compute_score as _compute_score
 
 __all__ = ["compute_score"]
