+++ tests/test_score_formula.py
@@ -1,6 +1,6 @@
 import math
 
-import agentic_index_cli.agentic_index as ai
+from agentic_index_cli import scoring as ai
 
 
 def test_compute_score_formula():
