+++ tests/test_search_pagination.py
@@ -1,4 +1,4 @@
-import agentic_index_cli.agentic_index as ai
+import agentic_index_cli.network as ai
 
 
 def test_search_and_harvest_pagination(monkeypatch):
